+++ rust/include/trajoptlib.h
@@ -54,8 +54,8 @@ class SwervePathBuilderImpl {
   void sgmt_polygon_obstacle(size_t from_idx, size_t to_idx,
                              rust::Vec<double> x, rust::Vec<double> y,
                              double radius);
-  HolonomicTrajectory generate(uint32_t handle) const;
-  void enable_state_feedback(rust::Fn<void(HolonomicTrajectory, uint32_t)> callback);
+  HolonomicTrajectory generate(int64_t handle) const;
+  void enable_state_feedback(rust::Fn<void(HolonomicTrajectory, int64_t)> callback);
   void cancel_all();
 
   SwervePathBuilderImpl() = default;
